Criei a pasta controller e criei o arquivo  UserControler e AccountController e coloquei o código do index para cada arquivo
